PLAT-104269: Rename AgateDecorator to ThemeDecorator (#244)
* copy AgateDecorator as ThemeDecorator

* deprecate AgateDecorator

* update sampler and ui tests

* remove AgateDecorator unit tests

* update version for deprecation

* move sampler AgateEnvironment to ThemeEnvironment

* remove duplicate (outdated) less and screenTypes
