+++ samples/sampler/src/ThemeEnvironment/ThemeEnvironment.js
@@ -7,13 +7,13 @@ import {color} from '@storybook/addon-knobs';
 import {Row, Column, Cell} from '@enact/ui/Layout';
 import {boolean, select} from '@enact/storybook-utils/addons/knobs';
 
-import AgateDecorator from '@enact/agate/AgateDecorator';
+import ThemeDecorator from '@enact/agate/ThemeDecorator';
 import Heading from '@enact/agate/Heading';
 import {Panels, Panel} from '@enact/agate/Panels';
 import Skinnable from '@enact/agate/Skinnable';
 import Scroller from '@enact/agate/Scroller';
 
-import css from './AgateEnvironment.module.less';
+import css from './ThemeEnvironment.module.less';
 
 const globalGroup = 'Global Knobs';
 
@@ -39,7 +39,7 @@ const reloadPage = () => {
 };
 
 const PanelsBase = kind({
-	name: 'AgateEnvironment',
+	name: 'ThemeEnvironment',
 
 	propTypes: {
 		description: PropTypes.string,
@@ -70,31 +70,31 @@ const PanelsBase = kind({
 });
 
 const FullscreenBase = kind({
-	name: 'AgateEnvironment',
+	name: 'ThemeEnvironment',
 
 	render: (props) => (
 		<div {...props} />
 	)
 });
 
-const Agate = AgateDecorator({overlay: false}, PanelsBase);
-const AgateFullscreen = AgateDecorator({overlay: false}, FullscreenBase);
+const Agate = ThemeDecorator({overlay: false}, PanelsBase);
+const AgateFullscreen = ThemeDecorator({overlay: false}, FullscreenBase);
 
 const locales = {
-	'local':                                                             '',
-	'en-US - US English':                                                'en-US',
-	'ko-KR - Korean':                                                    'ko-KR',
-	'es-ES - Spanish, with alternate weekends':                          'es-ES',
-	'am-ET - Amharic, 5 meridiems':                                      'am-ET',
-	'th-TH - Thai, with tall characters':                                'th-TH',
-	'ar-SA - Arabic, RTL and standard font':                             'ar-SA',
-	'ur-PK - Urdu, RTL and custom Urdu font':                            'ur-PK',
-	'zh-Hans-HK - Simplified Chinese, custom Hans font':                 'zh-Hans-HK',
-	'zh-Hant-HK - Traditional Chinese, custom Hant font':                'zh-Hant-HK',
-	'vi-VN - Vietnamese, Special non-latin font handling':               'vi-VN',
-	'ta-IN - Tamil, custom Indian font':                                 'ta-IN',
-	'ja-JP - Japanese, custom Japanese font':                            'ja-JP',
-	'en-JP - English, custom Japanese font':                             'en-JP',
+	'local': '',
+	'en-US - US English': 'en-US',
+	'ko-KR - Korean': 'ko-KR',
+	'es-ES - Spanish, with alternate weekends': 'es-ES',
+	'am-ET - Amharic, 5 meridiems': 'am-ET',
+	'th-TH - Thai, with tall characters': 'th-TH',
+	'ar-SA - Arabic, RTL and standard font': 'ar-SA',
+	'ur-PK - Urdu, RTL and custom Urdu font': 'ur-PK',
+	'zh-Hans-HK - Simplified Chinese, custom Hans font': 'zh-Hans-HK',
+	'zh-Hant-HK - Traditional Chinese, custom Hant font': 'zh-Hant-HK',
+	'vi-VN - Vietnamese, Special non-latin font handling': 'vi-VN',
+	'ta-IN - Tamil, custom Indian font': 'ta-IN',
+	'ja-JP - Japanese, custom Japanese font': 'ja-JP',
+	'en-JP - English, custom Japanese font': 'en-JP',
 	'si-LK - Sinhala, external font family with different line metrics': 'si-LK'
 };
 
@@ -116,7 +116,7 @@ const getPropFromURL = (propName, fallbackValue) => {
 	if (startIndex > -1) {
 		const keyIndex = locationParams.indexOf('=', startIndex);
 
-		if (locationParams.indexOf('&', keyIndex) > -1 ) {
+		if (locationParams.indexOf('&', keyIndex) > -1) {
 			const valueIndex = locationParams.indexOf('&', keyIndex);
 			return decodeURIComponent(locationParams.substring(keyIndex + 1, valueIndex));
 		} else {
