+++ samples/sampler/src/ThemeEnvironment/ThemeEnvironment.module.less
@@ -1,4 +1,4 @@
-// AgateEnvironment.module.less
+// ThemeEnvironment.module.less
 //
 @import "~../../../../styles/mixins.less";
 @import "~../../../../styles/skin.less";
@@ -8,7 +8,7 @@
 	position: relative;
 
 	.applySkins({
-		// These background rules, and the &::before below are cloned from AgateDecorator to mimic the style an app gets.
+		// These background rules, and the &::before below are cloned from ThemeDecorator to mimic the style an app gets.
 		background-color: @agate-bg-color;
 		background-image: @agate-bg-image;
