+++ samples/sampler/src/ThemeEnvironment/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "ThemeEnvironment.js"
+}
