+++ tests/ui/apps/RadioItem/Agate-RadioItem-View.js
@@ -1,5 +1,5 @@
 import RadioItem from '../../../../RadioItem';
-import AgateDecorator from '../../../../AgateDecorator';
+import ThemeDecorator from '../../../../ThemeDecorator';
 import React from 'react';
 
 const app = (props) => <div {...props}>
@@ -18,4 +18,4 @@ const app = (props) => <div {...props}>
 	</div>
 </div>;
 
-export default AgateDecorator(app);
+export default ThemeDecorator(app);
