+++ tests/ui/apps/ToggleButton/Agate-ToggleButton-View.js
@@ -1,5 +1,5 @@
 import ToggleButton from '../../../../ToggleButton';
-import AgateDecorator from '../../../../AgateDecorator';
+import ThemeDecorator from '../../../../ThemeDecorator';
 import React from 'react';
 
 const app = (props) => <div {...props}>
@@ -29,4 +29,4 @@ const app = (props) => <div {...props}>
 	</div>
 </div>;
 
-export default AgateDecorator(app);
+export default ThemeDecorator(app);
